requirements.txt
torch
torchvision
torchaudio
numpy
Pillow
scikit-learn
streamlit
matplotlib
